requirements.txt
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
